chore: merge build config for core package
