+++ packages/core/modern.config.ts
@@ -1,17 +1,60 @@
-import { defineConfig, moduleTools } from '@modern-js/module-tools';
+import fs from 'node:fs';
+import path from 'node:path';
 import {
-  BUILD_TARGET,
-  cjsBuildConfig,
-  commonExternals,
-  emitTypePkgJsonPlugin,
-  esmBuildConfig,
-} from '@rsbuild/config/modern.config.ts';
+  type CliPlugin,
+  type ModuleTools,
+  defineConfig,
+  moduleTools,
+} from '@modern-js/module-tools';
 import prebundleConfig from './prebundle.config.mjs';
 
+const define = {
+  RSBUILD_VERSION: require('../../packages/core/package.json').version,
+};
+
+const BUILD_TARGET = {
+  node: 'es2021',
+  client: 'es2017',
+} as const;
+
+const requireShim = {
+  // use import.meta['url'] to bypass bundle-require replacement of import.meta.url
+  js: `import { createRequire } from 'module';
+var require = createRequire(import.meta['url']);\n`,
+};
+
+const emitTypePkgJsonPlugin: CliPlugin<ModuleTools> = {
+  name: 'emit-type-pkg-json-plugin',
+
+  setup() {
+    return {
+      afterBuild() {
+        const typesDir = path.join(process.cwd(), 'dist-types');
+        const pkgPath = path.join(typesDir, 'package.json');
+        if (!fs.existsSync(typesDir)) {
+          fs.mkdirSync(typesDir);
+        }
+        fs.writeFileSync(
+          pkgPath,
+          JSON.stringify({
+            '//': 'This file is for making TypeScript work with moduleResolution node16+.',
+            version: '1.0.0',
+          }),
+          'utf8',
+        );
+      },
+    };
+  },
+};
+
 const externals = [
-  ...commonExternals,
+  'webpack',
+  '@rspack/core',
+  '@rsbuild/core',
   '@rsbuild/core/client/hmr',
   '@rsbuild/core/client/overlay',
+  /[\\/]compiled[\\/]/,
+  /node:/,
 ];
 
 const aliasCompiledPlugin = {
@@ -33,7 +76,12 @@ export default defineConfig({
   buildConfig: [
     // Node / ESM
     {
-      ...esmBuildConfig,
+      format: 'esm',
+      target: BUILD_TARGET.node,
+      define,
+      autoExtension: true,
+      shims: true,
+      banner: requireShim,
       input: ['src/index.ts'],
       externals,
       dts: false,
@@ -44,7 +92,11 @@ export default defineConfig({
     },
     // Node / CJS
     {
-      ...cjsBuildConfig,
+      format: 'cjs',
+      target: BUILD_TARGET.node,
+      define,
+      autoExtension: true,
+      dts: false,
       input: [
         'src/index.ts',
         'src/loader/ignoreCssLoader.ts',
